# cmd/car-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "'This is a car api server'",
        "title": "car-api",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/car": {
            "delete": {
                "description": "delete car",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "delete car",
                "operationId": "delete-car",
                "parameters": [
                    {
                        "description": "reg num",
                        "name": "regNum",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/car.DeleteCarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/car.emptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "update car",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "update car",
                "operationId": "update-car",
                "parameters": [
                    {
                        "description": "car info for update",
                        "name": "CarInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/car.UpdateCarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/car.emptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/cars": {
            "get": {
                "description": "get cars",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "get cars",
                "operationId": "get-cars",
                "parameters": [
                    {
                        "description": "filter for car list",
                        "name": "Filter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/car.GetCarsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/car.GetCarsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "add cars",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "add cars",
                "operationId": "add-cars",
                "parameters": [
                    {
                        "description": "reg num list",
                        "name": "regNums",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/car.AddCarRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/car.emptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "car.AddCarRequest": {
            "type": "object",
            "properties": {
                "reg_nums": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "car.DeleteCarRequest": {
            "type": "object",
            "properties": {
                "reg_num": {
                    "type": "string"
                }
            }
        },
        "car.GetCarsRequest": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "mark": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/models.Person"
                },
                "page": {
                    "type": "integer"
                },
                "reg_num": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "car.GetCarsResponse": {
            "type": "object",
            "properties": {
                "cars": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CarInfo"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "car.UpdateCarRequest": {
            "type": "object",
            "properties": {
                "mark": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/models.Person"
                },
                "reg_num": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "car.emptyResponse": {
            "type": "object"
        },
        "models.CarInfo": {
            "type": "object",
            "properties": {
                "mark": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/models.Person"
                },
                "reg_num": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "models.Person": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        }
    }
}